added onboarding

+++ app/schemas/onboarding.py
@@ -56,7 +56,7 @@ class IncomeInfo(BaseModel):
 
 class LiabilitiesInfo(BaseModel):
 
-    credit_cards: float = Field(..., ge=0, description="Number of Credit cards")
+    credit_cards: int = Field(..., ge=0, description="Number of Credit cards")
     credit_cards_limit: float = Field(..., ge=0, description="Credit cards amount limit")
     auto_loan: float = Field(..., ge=0, description="Auto loan remaining amount")
     personal_loan: float = Field(..., ge=0, description="Personal loan remaining amount")
